Removed legacy vb functions and HiRes support (since it was slowing down the emulator and just a few games uses it), also other changes

+++ SNES.net/Renderer/IRenderer.vb
@@ -1,5 +1,5 @@
 ﻿Public Interface IRenderer
-    Sub RenderBuffer(Buffer() As Byte)
+    Sub RenderBuffer(Buffer() As Integer)
     Sub SetTargetControl(Control As PictureBox)
     Sub SetZoom(Size As Integer)
